build toolchain only
if env var is set,
- skip psl1ght
- skip ps3libraries

+++ scripts/008-psl1ght.sh
@@ -4,6 +4,12 @@
 ## Download the source code.
 wget --no-check-certificate https://github.com/ps3dev/PSL1GHT/tarball/master -O psl1ght.tar.gz
 
+## Check if we want to skip this step
+if [[ -n "$BUILD_PS3TOOLCHAIN_ONLY" ]]; then
+    echo "PS3 Toolchain only set. Skipping..."
+    exit 0
+fi
+
 ## Unpack the source code.
 rm -Rf psl1ght && mkdir psl1ght && tar --strip-components=1 --directory=psl1ght -xvzf psl1ght.tar.gz
